Add html main tag skeleton
This commit adds the html elements for the main tag, which includes the
speedtest, results, also and navigator classes for each element present
in a normal google results webpage. Sub-classes were added for special
elements such as an Ad notification or the far edges of the Gooooooogle
navigation bottom bar.