Add styling instructions and fix minor typo
This commit adds the docs folder alongside a styling.md file which
describes the process taken to style the webpage from its html
skeleton. Along with this commit a simple README.md typo fix was added
since it's too minor to justify a single commit, consider this to be a
"documentation" oriented commit.