Update index html to include viewboxes for navbot
Opted for the viewBox approach when handling google's navigation bar
icons given its flexibility and consistency when applied to multiple
elements.

This commit also removes an element in the html that is no longer
present in google's current results page, and renames a css class to be
more in line with its sibling elements.